Break out PageLimitingOctokitClient tests
