+++ test/services/download_service_test.rb
@@ -94,62 +94,4 @@ class DownloadServiceTest < ActiveSupport::TestCase
     )
     download_service.download
   end
-
-  test 'fetch_notifications pages until the end' do
-    setup_paging_stubs
-    download_service = DownloadService.new(create(:morty))
-    fetched_notifications = download_service.fetch_notifications(params: {per_page: 2}, max_results:50)
-    assert_equal 8, fetched_notifications.size
-  end
-
-  test 'fetch_notifications pages stops at max_notifications' do
-    setup_paging_stubs
-    download_service = DownloadService.new(create(:morty))
-    fetched_notifications = download_service.fetch_notifications(params: {per_page: 2}, max_results:5)
-    assert_equal 5, fetched_notifications.size
-  end
-
-  test 'fetch_notifications respects Octobox.config.max_notifications_to_sync' do
-    setup_paging_stubs
-    Octobox.config.stubs(:max_notifications_to_sync).returns(5)
-    download_service = DownloadService.new(create(:morty))
-    fetched_notifications = download_service.fetch_notifications(params: {per_page: 2})
-    assert_equal 5, fetched_notifications.size
-  end
-
-  def setup_paging_stubs
-    [
-      {
-        page_number: 1,
-        link_header: '<https://api.github.com/notifications?page=2&per_page=2>; rel="next"'
-      },
-      {
-        page_number: 2,
-        link_header: '<https://api.github.com/notifications?page=3&per_page=2>; rel="next"'
-      },
-      {
-        page_number: 3,
-        link_header: '<https://api.github.com/notifications?page=4&per_page=2>; rel="next"'
-      },
-      {
-        page_number: 4,
-        link_header: '<https://api.github.com/notifications?page=1&per_page=2>; rel="first"'
-      }
-    ].each do |page|
-      notifications = JSON.parse(file_fixture('morty_notifications.json').read)
-      id = page[:page_number] * 2
-      notifications.each do |n|
-        n['id'] = id.to_s
-        id += 1
-      end
-      page_url = page[:page_number] > 1 ?
-        "https://api.github.com/notifications?page=#{page[:page_number]}&per_page=2" :
-        'https://api.github.com/notifications?per_page=2'
-      stub_notifications_request(
-        url: page_url,
-        body: notifications.to_json,
-        extra_headers: {'Link' => page[:link_header]}
-      )
-    end
-  end
 end
